Fixed more warnings from cdash
Specifically, fixes:

/Common/DataModel/vtkDataSet.cxx:1075:44: warning: constructor parameter 'BitSet' shadows the field 'BitSet' of 'IsAnyBitSetFunctor' [-Wshadow-field-in-constructor]
  IsAnyBitSetFunctor(vtkUnsignedCharArray* BitSet, int BitFlag)
                                           ^
/Common/DataModel/vtkDataSet.cxx:1075:56: warning: constructor parameter 'BitFlag' shadows the field 'BitFlag' of 'IsAnyBitSetFunctor' [-Wshadow-field-in-constructor]
  IsAnyBitSetFunctor(vtkUnsignedCharArray* BitSet, int BitFlag)

/Filters/General/vtkFiniteElementFieldDistributor.cxx:819:43: warning: comparison of integers of different signs: 'std::vector<std::vector<double>>::size_type' (aka 'unsigned long') and 'const int' [-Wsign-compare]
        reqInit |= (*hCurlMats)[i].size() != npts;
                   ~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~
/Filters/General/vtkFiniteElementFieldDistributor.cxx:820:42: warning: comparison of integers of different signs: 'std::vector<std::vector<double>>::size_type' (aka 'unsigned long') and 'const int' [-Wsign-compare]
        reqInit |= (*hDivMats)[i].size() != npts;
                   ~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~

/Filters/General/vtkFiniteElementFieldDistributor.cxx:1513:16: warning: moving a temporary object prevents copy elision [-Wpessimizing-move]
      coeffs = std::move(::GetEdgeAttributes(fieldName, oldCd, cellId));
               ^

/Filters/General/vtkFiniteElementFieldDistributor.cxx:1517:16: warning: moving a temporary object prevents copy elision [-Wpessimizing-move]
      coeffs = std::move(::GetFaceAttributes(fieldName, oldCd, cellId));
               ^

/IO/IOSS/vtkIOSSReader.cxx:192:26: warning: unused function 'Split' [-Wunused-function]

/Charts/Core/vtkChartParallelCoordinates.cxx:230:27: warning: comparison of integers of different signs: 'int' and 'std::vector<std::vector<float>>::size_type' (aka 'unsigned long') [-Wsign-compare]

/IO/Cesium3DTiles/TreeInformation.cxx:613:92: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]

/IO/Cesium3DTiles/TreeInformation.cxx:659:21: warning: comparison of integers of different signs: 'int' and 'std::vector<std::vector<unsigned long>>::size_type' (aka 'unsigned long') [-Wsign-compare]

/IO/Cesium3DTiles/TreeInformation.cxx:661:23: warning: comparison of integers of different signs: 'int' and 'std::vector<unsigned long>::size_type' (aka 'unsigned long') [-Wsign-compare]

/Rendering/Volume/Testing/Cxx/TestGPURayCastMapperShadows.cxx:46:20: warning: unused function 'idxToCoords' [-Wunused-function]

/Filters/Geometry/vtkGeometryFilter.cxx:2104:7: warning: field 'Strips' will be initialized after field 'VertsConnPtr' [-Wreorder-ctor]

/Charts/Core/vtkChartParallelCoordinates.cxx:234:23: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]

/Common/DataModel/vtkCellTreeLocator.cxx:315:9: warning: field 'Cnt' will be initialized after field 'Min' [-Wreorder-ctor]

/Common/DataModel/vtkCellTreeLocator.cxx:1227:22: warning: comparison of integers of different signs: 'int' and 'std::vector<std::pair<vtkBoundingBox, int>>::size_type' (aka 'unsigned long') [-Wsign-compare]

/IO/IOSS/vtkIOSSWriter.cxx:62:18: warning: private field 'Self' is not used [-Wunused-private-field]

/IO/IOSS/vtkIOSSModel.cxx:780:24: warning: loop variable 'tuple' binds to a temporary value produced by a range of type 'typename detail::SelectTupleRange<vtkIntArray *, 0>::type' (aka 'vtk::detail::TupleRange<vtkAOSDataArrayTemplate<int>, 0>') [-Wrange-loop-bind-reference]

+++ Charts/Core/vtkChartParallelCoordinates.cxx
@@ -227,17 +227,17 @@ bool vtkChartParallelCoordinates::Paint(vtkContext2D* painter)
   }
 
   // Draw all stored range
-  for (int axis = 0; axis < this->Storage->AxesSelections.size(); ++axis)
+  for (size_t axis = 0; axis < this->Storage->AxesSelections.size(); ++axis)
   {
     size_t size = this->Storage->AxesSelections[axis].size();
     size = size - size % 2;
-    for (int j = 0; j < size; j += 2)
+    for (size_t j = 0; j < size; j += 2)
     {
       float min = this->Storage->AxesSelections[axis][j];
       float max = this->Storage->AxesSelections[axis][j + 1];
       if (min != max)
       {
-        this->PaintRect(painter, axis, min, max);
+        this->PaintRect(painter, static_cast<int>(axis), min, max);
       }
     }
   }
@@ -734,7 +734,7 @@ void vtkChartParallelCoordinates::ResetSelection()
 //------------------------------------------------------------------------------
 void vtkChartParallelCoordinates::ResetAxeSelection(int axe)
 {
-  for (int i = 0; i < this->Storage->AxesSelections[axe].size(); i += 2)
+  for (size_t i = 0; i < this->Storage->AxesSelections[axe].size(); i += 2)
   {
     float& min = this->Storage->AxesSelections[axe][i];
     float& max = this->Storage->AxesSelections[axe][i + 1];
@@ -794,7 +794,7 @@ void vtkChartParallelCoordinates::UpdateCurrentAxisSelection(int axisId)
   size_t size = this->Storage->AxesSelections[axisId].size();
   size =
     this->Storage->AxesSelections[axisId].size() - this->Storage->AxesSelections[axisId].size() % 2;
-  for (int i = 0; i < size; i += 2)
+  for (size_t i = 0; i < size; i += 2)
   {
     if (this->GetSelectionMode() == vtkContextScene::SELECTION_TOGGLE)
     {
